Finally fixed logout and SSO problems (I hope).
The "continue" address parameter of the logout url had a different subdomain to the calling page. This caused a redirect which added a slash to the url, and this in turn caused the 500 error.

SSO via cookie manipulation was a matter of trial and error - but now a user is logged into the site across all namespaces.